LayoutTests/imported/w3c/web-platform-tests/svg/animations/svglength-animation-invalid-value-5-expected.txt


FAIL Test SVGLength animation with invalid value: No spaces between number and unit in to attribute. assert_equals: expected 100 but got 50
